AMpy/requirements.txt
flask
pandas
psycopg2-binary
confluent-kafka
python-dotenv
